* 'java' of https://github.com/mar0x/unit:
  Improving Java module compile parameters detection.
  Introducing system jars class loader.
  Initializing classpath array with NULL.
  Terminating modules path with 0.
  Adjusting types to avoid compiler warning/error.
  Implementing POST body parsing to get parameters.
  Mapping filters to URL pattern and servlet.
  PHP: fixed "disable_functions" and "disable_classes" options.
  Packages: added strict version dependency for unit-dev/unit-devel.

+++ src/java/nxt_jni_InputStream.c
@@ -214,7 +214,7 @@ nxt_java_InputStream_skip(JNIEnv *env, jclass cls, jlong req_info_ptr, jlong n)
         buf = nxt_unit_buf_next(buf);
     }
 
-    n = n < (jlong) req->content_length ? n : req->content_length;
+    n = n < (jlong) req->content_length ? n : (jlong) req->content_length;
 
     req->content_length -= n;
